+++ src/modules/categories/categories.module.ts
@@ -1,8 +1,8 @@
 import { Module } from '@nestjs/common';
 
 import { EntitiesModule } from 'src/entities/entities.module';
-import { CategoriesController } from './categories.controller';
-import { CategoriesService } from './categories.service';
+import { CategoriesController } from './controllers/categories.controller';
+import { CategoriesService } from './services/categories.service';
 
 
 @Module({
